+++ src/model/visible/Visible.java
@@ -7,7 +7,7 @@ import model.turtle.Turtle;
  *
  */
 public abstract class Visible {
-	protected int show(Turtle t, boolean isVisible) {
+	protected double show(Turtle t, boolean isVisible) {
 		t.setShowing(isVisible);
 		t.getImageView().setVisible(isVisible);
 		return isVisible ? 1 : 0;
